requirements.txt
-e git+ssh://git@github.com/Molmed/allium.git@0eab95332794cfe1723a5dfb0d4d27b2069cfc69#egg=allium
-e git+ssh://git@github.com/Molmed/conformist.git@a3768859762e05fd6937d2fbb7d19f64d9f46df9#egg=conformist
contourpy==1.1.1
cycler==0.12.1
fonttools==4.53.1
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyyml==0.0.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
UpSetPlot==0.9.0
zipp==3.20.2
